1. Добавил 1 файл со схемой в файл contracts.py 2. Переименовал файл проверки (test_resource_get.py в формате snake_case
