+++ Schema/contracts.py
@@ -9,3 +9,15 @@ USER_DATA_SCHEME = {
         },
     "required": ["id", "email", "first_name", "last_name", "avatar"]
 }
+
+LIST_RESOURCE_SCHEME = {
+    "type": "object",
+    "properties": {
+        "id": {"type": "number"},
+        "name": {"type": "string"},
+        "year": {"type": "number"},
+        "color": {"type": "string"},
+        "pantone_value": {"type": "string"}
+        },
+    "required": ["id", "name", "year", "color", "pantone_value"]
+}
